it's a big push haha

+++ clientSuiviMarches/src/main/java/ma/ensak/clientSuiviMarches/controller/EmployeeController.java
@@ -3,6 +3,7 @@ package ma.ensak.clientSuiviMarches.controller;
 import java.util.Map;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -13,9 +14,12 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+
+
 import ma.ensak.clientSuiviMarches.beans.Employee;
 import ma.ensak.clientSuiviMarches.beans.Service;
 import ma.ensak.clientSuiviMarches.proxies.MicroserviceEmployeeProxy;
+import ma.ensak.clientSuiviMarches.roles.Roles;
 
 
 
@@ -30,6 +34,7 @@ public class EmployeeController {
 	@RequestMapping(value="addOrUpdateEmployeeRedirect")
 	public String addOrUpdateEmployee(HttpServletRequest request , Model model ,@RequestParam(defaultValue="")String message ,@RequestParam(defaultValue="") String idd  , Employee employee)
 	{
+		if(!Roles.isALL(request))  return "redirect:/authentification";
 		
 		model.addAttribute("message", "".equals(message) ? "" : message );
 		model.addAttribute("Sectiontitle", "Employees");
@@ -41,6 +46,13 @@ public class EmployeeController {
 	@RequestMapping(value="addOrUpdateEmployee", method=RequestMethod.POST)
 	public ModelAndView addOrUpdateService( HttpServletRequest request  ,RedirectAttributes redirectAttributes, @RequestParam Map<String, String> params)
 	{
+		ModelAndView modelAndView = null ; 
+		
+		if(!Roles.isALL(request))  {
+			modelAndView = new ModelAndView("redirect:/authentification");
+			return modelAndView;
+		}
+		
 		Employee employee=new Employee();
 		
 		if(!params.get("id").toString().equals("")) employee.setId(Long.parseLong(params.get("id")));
@@ -52,7 +64,7 @@ public class EmployeeController {
 		if(params.get("service") != null) employee.setService(microSEProxy.getServiceById(Long.parseLong(params.get("service"))).get());
 
 		Long id = employee.getId();
-		ModelAndView modelAndView = null ;
+		
 		if(id ==null )
 		{
 			microSEProxy.createEmployee(employee);
@@ -68,6 +80,10 @@ public class EmployeeController {
 	@RequestMapping(value="deleteEmployee")
 	public ModelAndView deleteEmployee( HttpServletRequest request , @RequestParam(defaultValue="")String idd)
 	{
+		if(!Roles.isALL(request))  {
+			return new ModelAndView("redirect:/authentification");
+			 
+		}
 		
 		microSEProxy.deleteEmployee(Long.parseLong(idd));;;
 		return new ModelAndView("redirect:/getEmployees","message","Suppréssion réussie");
@@ -77,9 +93,50 @@ public class EmployeeController {
 	public String getServices( HttpServletRequest request , Model model,@RequestParam(defaultValue="") String message)
 	{
 		
+		if(!Roles.isALL(request))  return "redirect:/authentification";
+		
 		model.addAttribute("message", "".equals(message) ? "" : message );
 		model.addAttribute("employees", microSEProxy.getAllEmployees());
 		model.addAttribute("Sectiontitle", "Employees");
 		return "employee/getEmployees";
+	}
+	
+	@RequestMapping(value="ProfileRedirect" , method=RequestMethod.GET)
+	public String profileRedirect( HttpServletRequest request, Model model )
+	{
+		if(!Roles.isALL(request)) return "redirect:/authentification";
+		
+		
+		Employee userSession = (Employee) request.getSession(false).getAttribute("userSession");
+			
+		userSession.setPassword(null);
+		model.addAttribute("Sectiontitle", "Profile");
+		model.addAttribute("employee", userSession);
+		return "employee/profile";
+	}
+	
+	@RequestMapping(value="UpdateProfile" , method=RequestMethod.POST)
+	public ModelAndView updateProfile( HttpServletRequest request, Model model , @RequestParam Map<String, String> params , RedirectAttributes redirectAttributes)
+	{
+		if(Roles.isALL(request)) return new ModelAndView("redirect:/authentification");
+		ModelAndView modelAndView = null ;
+        Employee employee=microSEProxy.getEmployeeById(Long.parseLong(params.get("id"))).get();
+		
+		employee.setEmail(params.get("email"));
+		employee.setPassword(params.get("password"));
+		
+		if (microSEProxy.createEmployee(employee)!=null)
+			{
+				modelAndView = new ModelAndView("redirect:ProfileRedirect", "message","Profile modifé");
+				HttpSession session = request.getSession(false);
+				 session.setAttribute("userSession", employee);
+			}
+		else
+		{
+			redirectAttributes.addFlashAttribute("employee",employee);
+			modelAndView = new ModelAndView("redirect:ProfileRedirect", "message","Login existe déjà");
+		}
+		
+		return modelAndView;
 	}
 }
